+++ app/src/main/java/com/quare/blitzsplit/MainActivity.kt
@@ -1,40 +1,22 @@
 package com.quare.blitzsplit
 
-import android.app.Activity
-import android.content.ContentValues.TAG
-import android.content.Context
 import android.os.Bundle
-import android.util.Log
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.material3.Button
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Surface
 import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.unit.dp
-import androidx.compose.ui.viewinterop.AndroidView
-import com.google.android.gms.ads.AdRequest
-import com.google.android.gms.ads.AdSize
-import com.google.android.gms.ads.AdView
-import com.google.android.gms.ads.LoadAdError
 import com.google.android.gms.ads.MobileAds
-import com.google.android.gms.ads.interstitial.InterstitialAd
-import com.google.android.gms.ads.interstitial.InterstitialAdLoadCallback
-import com.google.android.gms.ads.rewarded.RewardedAd
-import com.google.android.gms.ads.rewarded.RewardedAdLoadCallback
+import com.quare.blitzsplit.ads.BannerAd
 import com.quare.blitzsplit.ads.InterstitialAdClass
-import com.quare.blitzsplit.ads.PremiumAd
 import com.quare.blitzsplit.ui.theme.BlitzSplitTheme
 
 class MainActivity : ComponentActivity() {
@@ -42,7 +24,7 @@ class MainActivity : ComponentActivity() {
         super.onCreate(savedInstanceState)
         setContent {
             val interstitialAd = InterstitialAdClass()
-            val premiumAd = PremiumAd()
+            val rewardedAd = com.quare.blitzsplit.ads.RewardedAd()
 
             MobileAds.initialize(this)
             BlitzSplitTheme {
@@ -61,7 +43,7 @@ class MainActivity : ComponentActivity() {
                         Button(onClick = { interstitialAd.showInterstitialAd(this@MainActivity) }) {
                             Text("Show interstitial ad")
                         }
-                        Button(onClick = { premiumAd.showPremiumAd(this@MainActivity) }) {
+                        Button(onClick = { rewardedAd.showRewardedAd(this@MainActivity) }) {
                             Text("Show premium ad")
                         }
                     }
@@ -69,18 +51,4 @@ class MainActivity : ComponentActivity() {
             }
         }
     }
-}
-
-@Composable
-fun BannerAd(adId: String) {
-    AndroidView(
-        modifier = Modifier.fillMaxWidth(),
-        factory = { context ->
-            AdView(context).apply {
-                setAdSize(AdSize.BANNER)
-                adUnitId = adId
-                loadAd(AdRequest.Builder().build())
-            }
-        }
-    )
 }
