+++ app/src/main/java/com/quare/blitzsplit/ads/BannerAd.kt
@@ -1,0 +1,23 @@
+package com.quare.blitzsplit.ads
+
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.viewinterop.AndroidView
+import com.google.android.gms.ads.AdRequest
+import com.google.android.gms.ads.AdSize
+import com.google.android.gms.ads.AdView
+
+@Composable
+fun BannerAd(adId: String) {
+    AndroidView(
+        modifier = Modifier.fillMaxWidth(),
+        factory = { context ->
+            AdView(context).apply {
+                setAdSize(AdSize.BANNER)
+                adUnitId = adId
+                loadAd(AdRequest.Builder().build())
+            }
+        }
+    )
+}
