+++ app/src/main/java/com/quare/blitzsplit/ads/NativeAd.kt
@@ -1,0 +1,11 @@
+package com.quare.blitzsplit.ads
+
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.viewinterop.AndroidView
+
+@Composable
+fun NativeAd(
+    adId: String
+) {
+
+}
